+++ torchmetrics/functional/classification/accuracy.py
@@ -16,8 +16,7 @@ from typing import Optional, Tuple
 import torch
 from torch import Tensor, tensor
 
-from torchmetrics.classification.stat_scores import _reduce_stat_scores
-from torchmetrics.functional.classification.stat_scores import _stat_scores_update
+from torchmetrics.functional.classification.stat_scores import _reduce_stat_scores, _stat_scores_update
 from torchmetrics.utilities.checks import _check_classification_inputs, _input_format_classification, _input_squeeze
 from torchmetrics.utilities.enums import AverageMethod, DataType, MDMCAverageMethod
